Notes DELETE route

+++ routes/notes.js
@@ -62,6 +62,17 @@ router.put("/:id", async (req, res) => {
 // DELETE /api/notes/:id - Delete a note
 router.delete("/:id", async (req, res) => {
   try {
+    const noteToDelete = await Note.findById(req.params.id);
+
+    if (!noteToDelete) {
+      return res.status(404).json({ message: "No note found with this id!" });
+    }
+
+    if (noteToDelete.user.toString() !== req.user._id.toString) {
+      return res
+        .status(403)
+        .json({ message: "User is not authorized to delete this note." });
+    }
     // This needs an authorization check
     const note = await Note.findByIdAndDelete(req.params.id);
     if (!note) {
